Adding .Rmd files and .csv datasets
Adding two versions of .Rmd files; one is the completed version of the capstone project including the questions, code, plots, and methodology/interpretation. The other .Rmd file is formatted to knit the plots of the project as a dashboard. 
Additionally, I'm adding two .csv dataset files that I use for analysis.